Make correct BSC <-> distconf interoperation protocol when enabling/disabling distconf
