+++ ydb/core/blobstorage/nodewarden/ya.make
@@ -10,7 +10,11 @@ SRCS(
     distconf_dynamic.cpp
     distconf_generate.cpp
     distconf_fsm.cpp
-    distconf_invoke.cpp
+    distconf_invoke.h
+    distconf_invoke_common.cpp
+    distconf_invoke_state_storage.cpp
+    distconf_invoke_static_group.cpp
+    distconf_invoke_storage_config.cpp
     distconf_mon.cpp
     distconf_persistent_storage.cpp
     distconf_scatter_gather.cpp
