+++ ydb/core/mind/bscontroller/commit_config.cpp
@@ -13,6 +13,8 @@ namespace NKikimr::NBsController {
         std::optional<TYamlConfig> YamlConfig;
         std::optional<std::optional<TString>> StorageYamlConfig;
         std::optional<NKikimrBlobStorage::TStorageConfig> StorageConfig;
+        std::optional<ui64> ExpectedStorageYamlConfigVersion;
+        std::unique_ptr<IEventHandle> Handle;
 
         ui64 GenerationOnStart = 0;
         TString FingerprintOnStart;
@@ -20,11 +22,14 @@ namespace NKikimr::NBsController {
     public:
         TTxCommitConfig(TBlobStorageController *controller, std::optional<TYamlConfig>&& yamlConfig,
                 std::optional<std::optional<TString>>&& storageYamlConfig,
-                std::optional<NKikimrBlobStorage::TStorageConfig>&& storageConfig)
+                std::optional<NKikimrBlobStorage::TStorageConfig>&& storageConfig,
+                std::optional<ui64> expectedStorageYamlConfigVersion, std::unique_ptr<IEventHandle> handle)
             : TTransactionBase(controller)
             , YamlConfig(std::move(yamlConfig))
             , StorageYamlConfig(std::move(storageYamlConfig))
             , StorageConfig(std::move(storageConfig))
+            , ExpectedStorageYamlConfigVersion(expectedStorageYamlConfigVersion)
+            , Handle(std::move(handle))
         {}
 
         TTxType GetTxType() const override { return NBlobStorageController::TXTYPE_COMMIT_CONFIG; }
@@ -44,6 +49,9 @@ namespace NKikimr::NBsController {
                 } else {
                     row.UpdateToNull<Schema::State::StorageYamlConfig>();
                 }
+            }
+            if (ExpectedStorageYamlConfigVersion) {
+                row.Update<Schema::State::ExpectedStorageYamlConfigVersion>(*ExpectedStorageYamlConfigVersion);
             }
             return true;
         }
@@ -75,10 +83,13 @@ namespace NKikimr::NBsController {
                 const bool hadStorageConfigBefore = Self->StorageYamlConfig.has_value();
 
                 Self->StorageYamlConfig = std::move(*StorageYamlConfig);
-                Self->StorageYamlConfigVersion = NYamlConfig::GetStorageMetadata(*Self->StorageYamlConfig).Version.value_or(0);
-                Self->StorageYamlConfigHash = NYaml::GetConfigHash(*Self->StorageYamlConfig);
+                Self->StorageYamlConfigVersion = 0;
+                Self->StorageYamlConfigHash = 0;
 
                 if (Self->StorageYamlConfig) {
+                    Self->StorageYamlConfigVersion = NYamlConfig::GetStorageMetadata(*Self->StorageYamlConfig).Version.value_or(0);
+                    Self->StorageYamlConfigHash = NYaml::GetConfigHash(*Self->StorageYamlConfig);
+
                     if (!update) {
                         update.emplace();
                     }
@@ -87,18 +98,25 @@ namespace NKikimr::NBsController {
                     update.emplace(); // issue an update without storage yaml version meaning we are in single-config mode
                 }
             }
+            if (ExpectedStorageYamlConfigVersion) {
+                Self->ExpectedStorageYamlConfigVersion = *ExpectedStorageYamlConfigVersion;
+            }
             if (update && Self->StorageYamlConfig) {
                 update->SetStorageConfigVersion(NYamlConfig::GetStorageMetadata(*Self->StorageYamlConfig).Version.value_or(0));
             }
 
-            Self->ConsoleInteraction->OnConfigCommit();
+            if (Handle) {
+                TActivationContext::Send(Handle.release());
+            } else {
+                Self->ConsoleInteraction->OnConfigCommit();
 
-            if (update) {
-                for (auto& node: Self->Nodes) {
-                    if (node.second.ConnectedServerId) {
-                        auto configPersistEv = std::make_unique<TEvBlobStorage::TEvControllerNodeServiceSetUpdate>();
-                        configPersistEv->Record.MutableYamlConfig()->CopyFrom(*update);
-                        Self->SendToWarden(node.first, std::move(configPersistEv), 0);
+                if (update) {
+                    for (auto& node: Self->Nodes) {
+                        if (node.second.ConnectedServerId) {
+                            auto configPersistEv = std::make_unique<TEvBlobStorage::TEvControllerNodeServiceSetUpdate>();
+                            configPersistEv->Record.MutableYamlConfig()->CopyFrom(*update);
+                            Self->SendToWarden(node.first, std::move(configPersistEv), 0);
+                        }
                     }
                 }
             }
@@ -107,8 +125,10 @@ namespace NKikimr::NBsController {
 
     ITransaction* TBlobStorageController::CreateTxCommitConfig(std::optional<TYamlConfig>&& yamlConfig,
             std::optional<std::optional<TString>>&& storageYamlConfig,
-            std::optional<NKikimrBlobStorage::TStorageConfig>&& storageConfig) {
-        return new TTxCommitConfig(this, std::move(yamlConfig), std::move(storageYamlConfig), std::move(storageConfig));
+            std::optional<NKikimrBlobStorage::TStorageConfig>&& storageConfig,
+            std::optional<ui64> expectedStorageYamlConfigVersion, std::unique_ptr<IEventHandle> handle) {
+        return new TTxCommitConfig(this, std::move(yamlConfig), std::move(storageYamlConfig), std::move(storageConfig),
+            expectedStorageYamlConfigVersion, std::move(handle));
     }
 
 } // namespace NKikimr::NBsController
